Further updates to weather/forecast issue
More changes.  Looking back, it seems like this was to help parse between sensor and weather type entities.  Weather would have forecast date/times built in while sensors would typically have a number suffix to chain them together.  My changes are probably going to break that.  So if you're reading this and you get errors on that section because you don't use "weather." and instead use multiple forecast sensors in series, this is probably not the fork you're looking for.